mark hypen rules files as having latin-1 encoding
Erlang R25 defaults to expecting utf-8 encoding, and will blow up if
this is not the case, and the file hasn't been marked as having a
different encoding.

This used to be a warning, but the deprecation has been removed, and it
just errors now.

Danish/Swedish/Norwegian characters in hypen files cause this to happen.